ovms_server_v2: change charge current to float
otherwise the decimals in the android app aren't of any use :)

+++ vehicle/OVMS.V3/components/ovms_server_v2/src/ovms_server_v2.cpp
@@ -1002,7 +1002,7 @@ void OvmsServerV2::TransmitMsgStat(bool always)
     << ","
     << StandardMetrics.ms_v_charge_voltage->AsInt()
     << ","
-    << StandardMetrics.ms_v_charge_current->AsInt()
+    << StandardMetrics.ms_v_charge_current->AsFloat()
     << ","
     << StandardMetrics.ms_v_charge_state->AsString("stopped")
     << ","
